+++ lib/src/test/java/io/github/mangila/ensure4j/BoolTest.java
@@ -26,12 +26,6 @@ public class BoolTest {
         assertThatThrownBy(() -> Ensure.isTrue(false, "test"))
                 .isInstanceOf(EnsureException.class)
                 .hasMessage("test");
-        assertThatThrownBy(() -> Ensure.isTrue(false, (Supplier<RuntimeException>) null))
-                .isInstanceOf(EnsureException.class)
-                .hasMessage("supplier was null");
-        assertThatThrownBy(() -> Ensure.isTrue(false, () -> null))
-                .isInstanceOf(EnsureException.class)
-                .hasMessage("supplier was given a null value");
     }
 
     @Test
@@ -50,12 +44,6 @@ public class BoolTest {
         assertThatThrownBy(() -> Ensure.isFalse(true, "test"))
                 .isInstanceOf(EnsureException.class)
                 .hasMessage("test");
-        assertThatThrownBy(() -> Ensure.isFalse(true, (Supplier<RuntimeException>) null))
-                .isInstanceOf(EnsureException.class)
-                .hasMessage("supplier was null");
-        assertThatThrownBy(() -> Ensure.isFalse(true, () -> null))
-                .isInstanceOf(EnsureException.class)
-                .hasMessage("supplier was given a null value");
     }
 
 }