+++ lib/src/test/java/io/github/mangila/ensure4j/NumbersTest.java
@@ -10,7 +10,7 @@ public class NumbersTest {
 
 
     @Test
-    @DisplayName("Happy path min()")
+    @DisplayName("Happy path min(int)")
     void min() {
         assertThatCode(() -> Ensure.min(-1, 0))
                 .doesNotThrowAnyException();
@@ -21,7 +21,7 @@ public class NumbersTest {
     }
 
     @Test
-    @DisplayName("Sad path min()")
+    @DisplayName("Sad path min(int)")
     void min1() {
         assertThatThrownBy(() -> Ensure.min(1, -10))
                 .isInstanceOf(EnsureException.class)
@@ -32,6 +32,31 @@ public class NumbersTest {
         assertThatThrownBy(() -> Ensure.min(1, -10, () -> new IllegalArgumentException("test value must be greater than or equal to 1")))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("test value must be greater than or equal to 1");
+    }
+
+    @Test
+    @DisplayName("Happy path max(int)")
+    void max() {
+        assertThatCode(() -> Ensure.max(100, -1))
+                .doesNotThrowAnyException();
+        assertThatCode(() -> Ensure.max(0, 0))
+                .doesNotThrowAnyException();
+        assertThatCode(() -> Ensure.max(1, -1))
+                .doesNotThrowAnyException();
+    }
+
+    @Test
+    @DisplayName("Sad path max(int)")
+    void max1() {
+        assertThatThrownBy(() -> Ensure.max(100, 101))
+                .isInstanceOf(EnsureException.class)
+                .hasMessage("value must be less than or equal to 100, but was 101");
+        assertThatThrownBy(() -> Ensure.max(100, 101, "test value must be less than or equal to 100"))
+                .isInstanceOf(EnsureException.class)
+                .hasMessage("test value must be less than or equal to 100");
+        assertThatThrownBy(() -> Ensure.max(100, 101, () -> new IllegalArgumentException("test value must be less than or equal to 100")))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessage("test value must be less than or equal to 100");
     }
 
 }